Added separate files for menu 1 bottom
Added separate files for menu 1 bottom. The "timer_and_database_bottom_JS.js" and "timer_and_database_bottom_PHP.php" that need to be separate because the tests will be different, and the results will be sent into different tables.

Menu 1 bottom now also had the front page "overlay" with a startbutton, the form, and the ability to send into the database. It is currently not possible to send the results into the database because of the MySQL database structure.